Refactor: Pulling Up Frequency and Probability
